feat: created taskController class with all method handling for task related requests
